# Better CUDA synchronization logic