Currently there are still more calls to `cudaDeviceSynchronize` than would be necessary. This PR improves the synchronization logic to reduce the number of those calls. In multi GPU settings synchronization is instead done via CUDA events. The performance boost turned out to be relatively small but it is measurable:

| GPU            | Model   | Test  | t/s master | t/s PR |
|----------------|---------|-------|------------|--------|
| RTX 3090       | 7b q4_0 | tg128 |      90.81 |  91.33 |
| P40 + GTX 1070 | 7b q4_0 | tg128 |      21.53 |  21.60 |

I also deduplicated the calls to `cudaGetLastError`.