@cmp-nct 2023-07-04
Just as a sidenote: I've attempted that modification as a branch on my glllm.cpp repo and the result was a 25% performance reduction on 4090+3090 combination with 1:1 tensorsplit surprisingly.
tk/sec went from low 50ies into high 30ies

Might not be relevant to llama.cpp ofc, the cuda versions are different but the syncro part is almost the same
I still use the multi stream, multi event version (But i tested it with and without that, same low performance with the new sync)